+++ src/Behaviours/Mutate/PrependValues.php
@@ -3,6 +3,7 @@
 namespace Somnambulist\Components\Collection\Behaviours\Mutate;
 
 use Somnambulist\Components\Collection\Contracts\Collection;
+use Somnambulist\Components\Collection\Utils\Value;
 use function array_unshift;
 
 /**
@@ -22,6 +23,8 @@ trait PrependValues
      */
     public function prepend(mixed ...$value): Collection|static
     {
+        Value::assertAllOfType($value, $this->type);
+
         array_unshift($this->items, ...$value);
 
         return $this;