+++ src/Behaviours/Mutate/ReplaceValues.php
@@ -3,6 +3,7 @@
 namespace Somnambulist\Components\Collection\Behaviours\Mutate;
 
 use Somnambulist\Components\Collection\Contracts\Collection;
+use Somnambulist\Components\Collection\Utils\Value;
 use function array_replace;
 use function array_replace_recursive;
 
@@ -21,6 +22,8 @@ trait ReplaceValues
      */
     public function replace(array ...$items): Collection|static
     {
+        Value::assertAllOfType($items, $this->type);
+
         $this->items = array_replace($this->items, ...$items);
 
         return $this;
@@ -35,6 +38,8 @@ trait ReplaceValues
      */
     public function replaceRecursively(array ...$items): Collection|static
     {
+        Value::assertAllOfType($items, $this->type);
+
         $this->items = array_replace_recursive($this->items, ...$items);
 
         return $this;